+++ cart.h
@@ -43,6 +43,6 @@ namespace Cart
 
         void LoadRom(std::string filename);
         void DumpHeaderToConsole();
-        void Write(uint8_t address, uint8_t data);
-        uint8_t Read(uint8_t address);
+        void Write(uint16_t address, uint8_t data);
+        uint8_t Read(uint16_t address);
 };
